Upload serverless demo function
Here we upload a serverless function that uses flask. If we add this function directly into AWS Lambda as a serverlessfunction, it would not execute. However, if we also uploaded flask.zip as an AWS lambda layer, require the new layer in the function, it would execute without failures.

This is to demonstrate that third party dependencies can be imported through these AWS Lambda Layers.